:sparkles: Separated out Controls from Camera in CameraScreen.kt
