+++ app/src/main/java/in/surajsau/jisho/ui/facenet/CameraScreen.kt
@@ -1,0 +1,34 @@
+package `in`.surajsau.jisho.ui.facenet
+
+import `in`.surajsau.jisho.ui.base.camera.Camera
+import `in`.surajsau.jisho.ui.base.camera.CameraAction
+import `in`.surajsau.jisho.ui.base.camera.CameraControls
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.*
+import androidx.compose.ui.Modifier
+
+@Composable
+fun CameraScreen(
+    modifier: Modifier,
+    onImageCaptured: (String) -> Unit,
+) {
+
+    var cameraAction by remember { mutableStateOf(CameraAction.None) }
+
+    Column(modifier = modifier) {
+
+        Camera(
+            modifier = Modifier
+                .fillMaxWidth()
+                .weight(1f),
+            cameraAction = cameraAction,
+            onImageCaptured = onImageCaptured
+        )
+
+        CameraControls(
+            modifier = Modifier.fillMaxWidth(),
+            onCameraAction = { cameraAction = it }
+        )
+    }
+}
